Fixed a bug in REMOVE_AUTH
CONTROL implies SELECT, INSERT, UPDATE, and DELETE. Thus when revoking
authorities in REMOVE_AUTH we must revoke CONTROL before revoking SELECT,
INSERT, UPDATE, or DELETE. Any other order results in SQL0558 (SQLSTATE 42504).

Furthermore, in the case that a user holds CONTROL authority on a table, and
also holds CONTROL on a view defined over that table, if CONTROL is revoked
from the underlying table, it is implicitly revoked from the view too. Any
subsequent attempt to revoke CONTROL from the view will result in SQL0556
(SQLSTATE 42504).

Given these two circumstances have the same SQLSTATE (and given we can only
trap SQLSTATEs, not SQLCODEs in stored procs) we handle the first with an ORDER
BY clause, and the second with a CONTINUE HANDLER (essentially ignoring it). It
would be better to handle the second with revocations in the correct order by
that involves ordering by a recursive query over TABDEP - possible, but complex
and expensive.